dvbaudiosink: turn off async state changes
When async state change is turned on and we are going
to paused state we are waiting for preroll and only then
state change happens.

Problem is that preroll sometimes doesn't happen at all
and we are stuck until FLUSH or EOS event happens.

To fix this we turn off async state changes, so we
will change state immediately without waiting for preroll

Only dvbaudiosink is not prerolling as described here:
http://forums.openpli.org/topic/38783-gstreamer10-pauseunpause-improvements/?view=findpost&p=504374

	modified:   gstdvbaudiosink.c