dvbmediasink: accept h264 early NALU caps (workaround)
Recent changes in h264parse forward caps without width and height and we have
to accept those caps in order to get later updated caps with more information.

Here is the description why that happens:

The problem here is that the AU delimiter appears in MPEG-TS streams before any
SPS/PPS, and needs to be forwarded. h264parse has no code to hold back the NALU
until caps have been determined (and implementing that right might be
problematic) - so it needs to set some caps and push the AU before it knows
anything more specific about the stream.

The caps might get updated again later if there are SEI packets that provide
more information again.

https://bugzilla.gnome.org/show_bug.cgi?id=755885

	modified:   gstdvbvideosink.c